+++ src/Home.js
@@ -1,9 +1,7 @@
 import './Home.css';
 import AddTopping from "./components/AddTopping";
 import ToppingDisplay from "./components/ToppingDisplay";
-import ToppingMenu from "./ToppingMenu";
-import AddPizza from "./AddPizza";
-import PizzaDisplay from "./PizzaDisplay";
+import ToppingMenu from "./components/ToppingMenu";
 
 import { useState, useEffect } from "react";
 
@@ -13,7 +11,6 @@ var selectedTopping = "";
 function Home() {
   const [filters, setFilters] = useState({});
   const [toppingData, setToppingData] = useState({ toppings: [] }); // Stores toppings
-  const [pizzaData, setPizzaData] = useState({ pizzas: [] }); // Stores pizzas
 
 
   useEffect(() => {
@@ -22,66 +19,67 @@ function Home() {
     .then((toppingData) => setToppingData({ toppings: toppingData }));
   }, []);
 
-  useEffect(() => {
-    fetch("http://localhost:3000/pizzas")
-    .then((response) => response.json())
-    .then((pizzaData) => setPizzaData({ pizzas: pizzaData }));
-  }, []);
+
 
   // function to return sliced array after slice
-  const hasDuplicate = (arrayObj, colName) => {
-    var hash = Object.create(null);
-    return arrayObj.some((arr) => {
-      return arr[colName] && (hash[arr[colName]] || !(hash[arr[colName]] = true));
-      });
-    };
 
 
 
 
   const addToppingToData = (topping) => {
-    console.log('here')
     let toppings = toppingData["toppings"]; // Stores state in data
     //topping.id = toppings.length; // sets id for each topping
+    console.log(toppings.length)
 
-    const requestOptions = {
-      method: "POST", // posts topping to db
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(topping) // converts topping to string
+    if (toppings.length >= 20) {
+      console.log("Topping length cannot exceed 20.")
     }
-
-    if (JSON.stringify(toppings).includes(JSON.stringify(topping.topping))) { // checks if 
-      console.log("In data");
-    }  
     else {
-      fetch("http://localhost:3000/toppings", requestOptions)
-      .then((response) => response.json())
-      .then((toppingData) => {
-        toppings.push(toppingData); // pushes item into this data array
-        setToppingData({ toppings: toppings }); // setData = currentData
-        console.log(toppings)
-      }); // *get[s] data from backend data
+
+      const requestOptions = {
+        method: "POST", // posts topping to db
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify(topping) // converts topping to string
+      }
+
+      if (JSON.stringify(toppings).includes(JSON.stringify(topping.topping))) { // checks if 
+        console.log("In data");
+      }  
+      else {
+        fetch("http://localhost:3000/toppings", requestOptions)
+        .then((response) => response.json())
+        .then((toppingData) => {
+          toppings.push(toppingData); // pushes item into this data array
+          setToppingData({ toppings: toppings }); // setData = currentData
+          console.log(toppings)
+        }); // *get[s] data from backend data
+      }
     }
   }
 
   const deleteTopping = (topping) => {
-    console.log(topping)
-    const toppings = toppingData["toppings"];
-    const requestOptions = {
-      method: "DELETE"
+    if (topping.id == 0) { // Don't delete first pizza,
+      console.log("Cannot delete 1st topping in data")
     }
-    fetch(`http://localhost:3000/toppings/${topping.id}`, requestOptions).then(
-      (response) => {
-        //Checks response code
-        if (response.ok) {
-          const idx = toppings.indexOf(topping); // finds index in array
-          toppings.splice(idx, 1);
-          setToppingData({toppings: toppings})
-        }
+    else {
+
+      const toppings = toppingData["toppings"];
+      const requestOptions = {
+        method: "DELETE"
       }
-    );
+      fetch(`http://localhost:3000/toppings/${topping.id}`, requestOptions).then(
+        (response) => {
+          //Checks response code
+          if (response.ok) {
+            const idx = toppings.indexOf(topping); // finds index in array
+            toppings.splice(idx, 1);
+            setToppingData({toppings: toppings})
+          }
+        }
+      );
+    }
   };
 
   // Used when we want to update a topping's name. Used in ToppingDisplay.js
@@ -108,123 +106,6 @@ function Home() {
       );
     }
   };
-
-
-
-  // PIZZAS
-
-  const addPizzaToData = (pizza) => {
-    let pizzas = pizzaData["pizzas"]; // Stores state in data
-
-    //  Checks for duplicate pizzas
-
-    if (JSON.stringify(pizzas).includes(JSON.stringify(pizza.toppings))) {
-      console.log("In data, pizza not addded");
-    }
-
-    // If no duplicates, add the pizza thru POST request
-
-    else {
-      const requestOptions = {
-        method: "POST", // posts topping to db
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(pizza) // converts topping to string
-      }
-
-      fetch("http://localhost:3000/pizzas", requestOptions)
-      .then((response) => response.json())
-      .then((pizzaData) => {
-        console.log(pizzaData)
-        if (JSON.stringify(pizzas).includes(JSON.stringify(pizzaData.toppings))) {
-          console.log("In data, pizza not added");
-        }
-        else {
-          pizzas.push(pizzaData); // pushes item into this data array
-          setPizzaData({ pizzas: pizzas }); // setData = currentData
-        }
-      }); 
-    }
-  }
-
-  const deletePizza = (pizza) => {
-    console.log(pizza)
-    const pizzas = pizzaData["pizzas"];
-    const requestOptions = {
-      method: "DELETE"
-    }
-    fetch(`http://localhost:3000/pizzas/${pizza.id}`, requestOptions).then(
-      (response) => {
-        //Checks response code
-        if (response.ok) {
-          const idx = pizzas.indexOf(pizza); // finds index in array
-          pizzas.splice(idx, 1);
-          setPizzaData({pizzas: pizzas})
-        }
-      }
-    );
-  };
-
-  // Used when we want to update a topping's name. Used in ToppingDisplay.js
-
-  const getSelectedTopping = (topping) => {
-    selectedTopping = topping.topping;
-    console.log(selectedTopping);
-    
-  };
-
-  // Updates pizza with global selected topping. This topping updates in getSelectedTopping function
-  const updatePizza = (pizza) => {
-    let pizzas = pizzaData["pizzas"];
-
-    // If the selected topping is in the pizza data already, delete it
-    if (JSON.stringify(pizza.toppings).includes(selectedTopping)) { // checks for duplicate topping in data
-      // check if another pizza has the same toppings, after delete. If duplicate pizza, do nothing
-      console.log("In data, delete topping");
-      
-      var idx = pizzas[pizza.id].toppings.indexOf(selectedTopping); // finds index in array
-      pizzas[pizza.id].toppings.splice(idx, 1)
-      setPizzaData({pizzas: pizzas})
-
-      var isDuplicate = hasDuplicate(pizzas, "toppings");
-      if (isDuplicate) {
-        console.log("Error: Duplicate Pizza")
-        pizzas[pizza.id].toppings.push(selectedTopping); // pushes item into this data array
-        setPizzaData({ pizzas: pizzas }); // setData = currentData
-      }
-        
-    }  
-
-    // Topping not in pizza data, add it
-    else {
-      console.log(pizzas[pizza.id])
-      pizzas[pizza.id].toppings.push(selectedTopping); // pushes item into this data array
-      setPizzaData({ pizzas: pizzas }); // setData = currentData
-
-      // Check for duplicate pizza
-      var isDuplicate = hasDuplicate(pizzas, "toppings");
-      if (isDuplicate) {
-        console.log("Error: Duplicate Pizza")
-        var idx = pizzas[pizza.id].toppings.indexOf(selectedTopping); // finds index in array
-        pizzas[pizza.id].toppings.splice(idx, 1)
-        setPizzaData({pizzas: pizzas})
-      }
-    }
-      const requestOptions = {
-        method: "PUT", // posts topping to db
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(pizza) // converts topping to string
-      }
-
-      fetch(`http://localhost:3000/pizzas/${pizza.id}`, requestOptions)
-      .then((response) => response.json())
-      .then((pizzaData) => {}); // *get[s] data from backend data
-    
-  };
-
 
 
 
@@ -255,7 +136,7 @@ function Home() {
     <div>
       <div className="title-content-box"> 
         <h1 className="title-text"> Welcome to your Pizza Store! </h1>
-        <h2 className="title-text"> Add a topping, update a topping, or delete one from the menu. Afterwards, time to make a pizza! </h2>
+        <h2 className="title-text"> Add a topping, update a topping, or delete one from the menu. Afterwards, go make a pizza! </h2>
         <a href="/chef" className="manage-pizza-link">
           Manage Pizzas
         </a>
